pblio: Cleanup of spc stats
