+++ apps/pblio/spc/spcstats.go
@@ -16,72 +16,9 @@
 package spc
 
 import (
-	"fmt"
-	"github.com/lpabon/goioworkload/spc1"
 	"github.com/lpabon/tm"
 	"time"
 )
-
-type IoStats struct {
-	Io      *spc1.Spc1Io
-	Start   time.Time
-	Latency time.Duration
-}
-
-// --------------------------------------------
-
-type IoMeter struct {
-	Latency tm.TimeDuration
-	Ios     uint64
-
-	// 4KB Blocks transferred
-	Blocks uint64
-}
-
-func (i *IoMeter) Collect(iostat *IoStats) {
-	i.Ios++
-	i.Blocks += uint64(iostat.Io.Blocks)
-	i.Latency.Add(iostat.Latency)
-}
-
-func (i *IoMeter) CsvDelta(prev *IoMeter, delta time.Duration) string {
-	return fmt.Sprintf("%v,"+ // Ios
-		"%v,"+ // Bytes Transferred
-		"%v,"+ // MB/s
-		"%v,", // Latency in usecs
-		i.Ios-prev.Ios,
-		(i.Blocks-prev.Blocks)*4*KB,
-		(float64((i.Blocks-prev.Blocks)*4*KB)/float64(MB))/delta.Seconds(),
-		i.Latency.DeltaMeanTimeUsecs(&prev.Latency))
-}
-
-// --------------------------------------------
-
-type AsuStats struct {
-	total, read, write IoMeter
-	Latency            tm.TimeDuration
-}
-
-func NewAsuStats() *AsuStats {
-	return &AsuStats{}
-}
-
-func (a *AsuStats) Collect(iostat *IoStats) {
-	a.total.Collect(iostat)
-	if iostat.Io.Isread {
-		a.read.Collect(iostat)
-	} else {
-		a.write.Collect(iostat)
-	}
-}
-
-func (a *AsuStats) CsvDelta(prev *AsuStats, delta time.Duration) string {
-	return a.read.CsvDelta(&prev.read, delta) +
-		a.write.CsvDelta(&prev.write, delta) +
-		a.total.CsvDelta(&prev.total, delta)
-}
-
-// --------------------------------------------
 
 type SpcStats struct {
 	total, read, write IoMeter
